requirements.txt
python-telegram-bot==20.7
APScheduler==3.10.4
python-dotenv==1.0.0
fastapi
uvicorn[standard]
python-multipart
